# docs/core/CORE_API_OPENAPI.yaml
openapi: 3.1.1
info:
  title: Вузы России — API ядра-бекенда
  version: 0.1.0
  description: 'Предложенный контракт MVP от 2026-09-24. Не описание текущей реализации. Только операции ядра, включая
    API, вызываемое backend сервисов. Доменные API других компонентов не входят. CORE_API_SPEC.md дополняет HTTP-контракт
    правилами JWT, изоляции, транзакций и bootstrap. Private операции имеют отдельный server и AND-требование machine
    + actor; на публичном ingress всегда 404. x-required-permission, x-required-machine-scope и x-network-access
    — документационные extensions, не автоматическое обеспечение доступа. Форматы JSON: snake_case, UUID string,
    UTC. Формат документа — OpenAPI 3.1.1 (https://spec.openapis.org/oas/v3.1.1.html).'
servers:
- url: https://core.example
  description: Публичный HTTPS origin ядра; заменить при развёртывании.
security:
- CoreBearer: []
...
paths:
...
  /api/v1/institution:
    get:
      tags:
      - Institutions
      summary: Получить свои вузы и назначенные профили
      operationId: listMyInstitutions
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstitutionPage'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '500':
          $ref: '#/components/responses/Error500'
      description: Только членства текущего пользователя. Пустой результат — 200; видны pending/active/suspended
        со статусом. Не глобальный каталог.
      parameters:
      - $ref: '#/components/parameters/LocaleQuery'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
  /api/v1/institution/{institution_id}:
    parameters:
    - $ref: '#/components/parameters/institution_id'
    get:
      tags:
      - Institutions
      summary: Получить доступный вуз
      operationId: getMyInstitution
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstitutionView'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - $ref: '#/components/parameters/LocaleQuery'
  /api/v1/institution/{institution_id}/profiles:
    parameters:
    - $ref: '#/components/parameters/institution_id'
    get:
      tags:
      - Institutions
      summary: Получить свои профили в вузе
      operationId: listMyProfiles
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileList'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '500':
          $ref: '#/components/responses/Error500'
  /api/v1/institution/{institution_id}/service:
    parameters:
    - $ref: '#/components/parameters/institution_id'
    get:
      tags:
      - Services
      summary: Получить сервисы и меню для выбранного профиля
      operationId: listMyServices
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServicePage'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      description: Только включённые экземпляры активного вуза с хотя бы одним доступным меню. Меню отфильтрованы
        по профилю и всем required_permissions.
      parameters:
      - $ref: '#/components/parameters/ProfileQuery'
      - $ref: '#/components/parameters/LocaleQuery'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
  /api/v1/institution/{institution_id}/service/{service_id}:
    parameters:
    - $ref: '#/components/parameters/institution_id'
    - $ref: '#/components/parameters/service_id'
    get:
      tags:
      - Services
      summary: Получить карточку экземпляра и адреса клиентов
      operationId: getMyService
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceView'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      description: service_id принадлежит institution_id. Только меню выбранного профиля. Неподдерживаемый сервисом
        профиль — 404, отсутствующий у пользователя профиль — 403, выключенный экземпляр — 409.
      parameters:
      - $ref: '#/components/parameters/ProfileQuery'
      - $ref: '#/components/parameters/LocaleQuery'
...
  /api/v1/institution/{institution_id}/groups:
    parameters:
    - $ref: '#/components/parameters/institution_id'
    get:
      tags:
      - Shell study groups
      summary: Учебные группы для оболочки
      operationId: listShellGroups
      parameters:
      - $ref: '#/components/parameters/ProfileQuery'
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShellGroupList'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '500':
          $ref: '#/components/responses/Error500'
      description: 'Студенту — только своя группа; преподавателю и администратору — все группы вуза с составом. can_manage и ETag — при праве groups.manage в администрировании.'
      x-extension: study-groups
      security:
      - CoreBearer: []
    post:
      tags:
      - Shell study groups
      summary: Создать группу (groups.manage)
      operationId: createShellGroup
      parameters:
      - $ref: '#/components/parameters/ProfileQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudyGroupInput'
      responses:
        '201':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudyGroup'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '500':
          $ref: '#/components/responses/Error500'
      description: 'Требует право groups.manage в administration для профиля admin.'
      x-extension: study-groups
      security:
      - CoreBearer: []
  /api/v1/institution/{institution_id}/groups/{group_id}:
    parameters:
    - $ref: '#/components/parameters/institution_id'
    - $ref: '#/components/parameters/group_id'
    patch:
      tags:
      - Shell study groups
      summary: Переименовать группу
      operationId: renameShellGroup
      parameters:
      - $ref: '#/components/parameters/ProfileQuery'
      - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudyGroupInput'
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudyGroup'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '500':
          $ref: '#/components/responses/Error500'
        '412':
          $ref: '#/components/responses/Error412'
        '428':
          $ref: '#/components/responses/Error428'
      description: 'Требует groups.manage и If-Match.'
      x-extension: study-groups
      security:
      - CoreBearer: []
    delete:
      tags:
      - Shell study groups
      summary: Удалить пустую группу
      operationId: deleteShellGroup
      parameters:
      - $ref: '#/components/parameters/ProfileQuery'
      - $ref: '#/components/parameters/IfMatch'
      responses:
        '204':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '500':
          $ref: '#/components/responses/Error500'
        '412':
          $ref: '#/components/responses/Error412'
        '428':
          $ref: '#/components/responses/Error428'
      description: 'Требует groups.manage и If-Match; непустая группа — 409.'
      x-extension: study-groups
      security:
      - CoreBearer: []
  /api/v1/institution/{institution_id}/groups/{group_id}/members:
    parameters:
    - $ref: '#/components/parameters/institution_id'
    - $ref: '#/components/parameters/group_id'
    put:
      tags:
      - Shell study groups
      summary: Заменить состав группы
      operationId: replaceShellGroupMembers
      parameters:
      - $ref: '#/components/parameters/ProfileQuery'
      - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudyGroupMembersInput'
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudyGroupMembers'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '500':
          $ref: '#/components/responses/Error500'
        '412':
          $ref: '#/components/responses/Error412'
        '428':
          $ref: '#/components/responses/Error428'
      description: 'Требует groups.manage и If-Match; студент — максимум в одной группе.'
      x-extension: study-groups
      security:
      - CoreBearer: []
...
  /api/v1/institution/{institution_id}/widgets:
    parameters:
    - $ref: '#/components/parameters/institution_id'
    get:
      tags:
      - Widgets
      summary: Виджеты главного экрана профиля
      operationId: listShellWidgets
      parameters:
      - $ref: '#/components/parameters/ProfileQuery'
      - $ref: '#/components/parameters/LocaleQuery'
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShellWidgetList'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '500':
          $ref: '#/components/responses/Error500'
      description: 'Только видимые профилю: включены сервисом и администратором вуза, есть все required_permissions; не больше 12 (WIDGETS_SPEC.md §6).'
      x-extension: widgets
      security:
      - CoreBearer: []
...
  /institution:
    get:
      tags:
      - Shell
      summary: Отдать HTML ядра-клиента
      operationId: getInstitutionShell
      responses:
        '200':
          description: HTML shell.
          content:
            text/html:
              schema:
                type: string
        '503':
          description: HTML оболочка временно недоступна.
      description: Публичная оболочка без персональных данных и JWT. Авторизация и redirect при одном вузе выполняются
        shell после JSON login. Не читает bearer из URL.
      security: []
  /institution/{institution_id}:
    parameters:
    - $ref: '#/components/parameters/institution_id'
    get:
      tags:
      - Shell
      summary: Отдать HTML ядра-клиента
      operationId: getInstitutionWorkspaceShell
      responses:
        '200':
          description: HTML shell.
          content:
            text/html:
              schema:
                type: string
        '503':
          description: HTML оболочка временно недоступна.
        '404':
          description: Некорректный UUID пути; данные/членство проверяются отдельно через JSON API.
      description: Публичная оболочка без персональных данных и JWT. Авторизация и redirect при одном вузе выполняются
        shell после JSON login. Не читает bearer из URL.
      security: []
components:
  securitySchemes:
    CoreBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Пользовательский token_use=core_access, aud=core-api; живая sid обязательна.
    MachineBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: token_use=machine_access, aud=core-internal. Credential и tenant проверяются на каждом запросе.
        Не пользовательский токен.
    ServiceBasic:
      type: http
      scheme: basic
      description: 'Только HTTPS token exchange: client_id как username, client_secret как password.'
    ActorToken:
      type: apiKey
      in: header
      name: X-Actor-Token
      description: 'JWT без префикса Bearer: service_access человека для экземпляра administration, профиль admin.
        AND с MachineBearer; актуальный RBAC из БД.'
...
  headers:
    RequestId:
      description: UUID запроса, назначенный ядром.
      schema:
        $ref: '#/components/schemas/UUID'
...